refactor: rename `core.py` -> `network.py`
