+++ pypsa/__init__.py
@@ -28,7 +28,7 @@ from pypsa import (
     plot,
     statistics,
 )
-from pypsa.core import Network, SubNetwork
+from pypsa.networks import Network, SubNetwork
 
 # e.g. "0.17.1" or "0.17.1.dev4+ga3890dc0" (if installed from git)
 __version__ = version("pypsa")
