+++ pypsa/descriptors.py
@@ -17,7 +17,7 @@ import pandas as pd
 from pypsa.utils import as_index, deprecated_common_kwargs
 
 if TYPE_CHECKING:
-    from pypsa.core import Network, SubNetwork
+    from pypsa.networks import Network, SubNetwork
 
 logger = logging.getLogger(__name__)
